requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
pytesseract==0.3.10
pyttsx3==2.90
Pillow==10.2.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1 